Review feedback round 2
- Move IpRange types to shared Nexus crate
- Remove more dead code
- Cleanup some comments
