+++ nexus/src/external_api/shared.rs
@@ -12,6 +12,9 @@ use serde::de::Error as _;
 use serde::Deserialize;
 use serde::Deserializer;
 use serde::Serialize;
+use std::net::IpAddr;
+use std::net::Ipv4Addr;
+use std::net::Ipv6Addr;
 use uuid::Uuid;
 
 /// Maximum number of role assignments allowed on any one resource
@@ -132,6 +135,167 @@ impl TryFrom<db::model::IdentityType> for IdentityType {
     }
 }
 
+/// An IP Range is a contiguous range of IP addresses, usually within an IP
+/// Pool.
+///
+/// The first address in the range is guaranteed to be no greater than the last
+/// address.
+#[derive(Clone, Copy, Debug, PartialEq, Deserialize, Serialize)]
+pub enum IpRange {
+    V4(Ipv4Range),
+    V6(Ipv6Range),
+}
+
+impl JsonSchema for IpRange {
+    fn schema_name() -> String {
+        "IpRange".to_string()
+    }
+
+    fn json_schema(
+        gen: &mut schemars::gen::SchemaGenerator,
+    ) -> schemars::schema::Schema {
+        schemars::schema::SchemaObject {
+            metadata: Some(
+                schemars::schema::Metadata { ..Default::default() }.into(),
+            ),
+            subschemas: Some(
+                schemars::schema::SubschemaValidation {
+                    one_of: Some(vec![
+                        omicron_common::api::external::label_schema(
+                            "v4",
+                            gen.subschema_for::<Ipv4Range>(),
+                        ),
+                        omicron_common::api::external::label_schema(
+                            "v6",
+                            gen.subschema_for::<Ipv6Range>(),
+                        ),
+                    ]),
+                    ..Default::default()
+                }
+                .into(),
+            ),
+            ..Default::default()
+        }
+        .into()
+    }
+}
+
+impl IpRange {
+    pub fn first_address(&self) -> IpAddr {
+        match self {
+            IpRange::V4(inner) => IpAddr::from(inner.first),
+            IpRange::V6(inner) => IpAddr::from(inner.first),
+        }
+    }
+
+    pub fn last_address(&self) -> IpAddr {
+        match self {
+            IpRange::V4(inner) => IpAddr::from(inner.last),
+            IpRange::V6(inner) => IpAddr::from(inner.last),
+        }
+    }
+}
+
+impl TryFrom<(Ipv4Addr, Ipv4Addr)> for IpRange {
+    type Error = String;
+
+    fn try_from(pair: (Ipv4Addr, Ipv4Addr)) -> Result<Self, Self::Error> {
+        Ipv4Range::new(pair.0, pair.1).map(IpRange::V4)
+    }
+}
+
+impl TryFrom<(Ipv6Addr, Ipv6Addr)> for IpRange {
+    type Error = String;
+
+    fn try_from(pair: (Ipv6Addr, Ipv6Addr)) -> Result<Self, Self::Error> {
+        Ipv6Range::new(pair.0, pair.1).map(IpRange::V6)
+    }
+}
+
+/// A non-decreasing IPv4 address range, inclusive of both ends.
+///
+/// The first address must be less than or equal to the last address.
+#[derive(Clone, Copy, Debug, PartialEq, Deserialize, Serialize, JsonSchema)]
+#[serde(try_from = "AnyIpv4Range")]
+pub struct Ipv4Range {
+    first: Ipv4Addr,
+    last: Ipv4Addr,
+}
+
+impl Ipv4Range {
+    pub fn new(first: Ipv4Addr, last: Ipv4Addr) -> Result<Self, String> {
+        if first <= last {
+            Ok(Self { first, last })
+        } else {
+            Err(String::from("IP address ranges must be non-decreasing"))
+        }
+    }
+
+    pub fn first_address(&self) -> Ipv4Addr {
+        self.first
+    }
+
+    pub fn last_address(&self) -> Ipv4Addr {
+        self.last
+    }
+}
+
+#[derive(Clone, Copy, Debug, Deserialize)]
+struct AnyIpv4Range {
+    first: Ipv4Addr,
+    last: Ipv4Addr,
+}
+
+impl TryFrom<AnyIpv4Range> for Ipv4Range {
+    type Error = Error;
+    fn try_from(r: AnyIpv4Range) -> Result<Self, Self::Error> {
+        Ipv4Range::new(r.first, r.last)
+            .map_err(|msg| Error::invalid_request(msg.as_str()))
+    }
+}
+
+/// A non-decreasing IPv6 address range, inclusive of both ends.
+///
+/// The first address must be less than or equal to the last address.
+#[derive(Clone, Copy, Debug, PartialEq, Deserialize, Serialize, JsonSchema)]
+#[serde(try_from = "AnyIpv6Range")]
+pub struct Ipv6Range {
+    first: Ipv6Addr,
+    last: Ipv6Addr,
+}
+
+impl Ipv6Range {
+    pub fn new(first: Ipv6Addr, last: Ipv6Addr) -> Result<Self, String> {
+        if first <= last {
+            Ok(Self { first, last })
+        } else {
+            Err(String::from("IP address ranges must be non-decreasing"))
+        }
+    }
+
+    pub fn first_address(&self) -> Ipv6Addr {
+        self.first
+    }
+
+    pub fn last_address(&self) -> Ipv6Addr {
+        self.last
+    }
+}
+
+#[derive(Clone, Copy, Debug, Deserialize)]
+struct AnyIpv6Range {
+    first: Ipv6Addr,
+    last: Ipv6Addr,
+}
+
+impl TryFrom<AnyIpv6Range> for Ipv6Range {
+    type Error = Error;
+    fn try_from(r: AnyIpv6Range) -> Result<Self, Self::Error> {
+        Ipv6Range::new(r.first, r.last)
+            .map_err(|msg| Error::invalid_request(msg.as_str()))
+    }
+}
+
 #[cfg(test)]
 mod test {
     use super::IdentityType;
@@ -139,10 +303,15 @@ mod test {
     use super::MAX_ROLE_ASSIGNMENTS_PER_RESOURCE;
     use crate::db;
     use crate::external_api::shared;
+    use crate::external_api::shared::IpRange;
+    use crate::external_api::shared::Ipv4Range;
+    use crate::external_api::shared::Ipv6Range;
     use anyhow::anyhow;
     use omicron_common::api::external::Error;
     use omicron_common::api::external::ResourceType;
     use serde::Deserialize;
+    use std::net::Ipv4Addr;
+    use std::net::Ipv6Addr;
 
     #[derive(Clone, Copy, Debug, Deserialize, Eq, PartialEq)]
     #[serde(rename_all = "kebab-case")]
@@ -264,5 +433,63 @@ mod test {
         assert_eq!(success.identity_type, IdentityType::SiloUser);
         assert_eq!(success.identity_id, identity_id);
         assert_eq!(success.role_name, DummyRoles::Bogus);
+    }
+
+    #[test]
+    fn test_ip_range_checks_non_decreasing() {
+        let lo = Ipv4Addr::new(10, 0, 0, 1);
+        let hi = Ipv4Addr::new(10, 0, 0, 3);
+        assert!(Ipv4Range::new(lo, hi).is_ok());
+        assert!(Ipv4Range::new(lo, lo).is_ok());
+        assert!(Ipv4Range::new(hi, lo).is_err());
+
+        let lo = Ipv6Addr::new(0xfd00, 0, 0, 0, 0, 0, 0, 1);
+        let hi = Ipv6Addr::new(0xfd00, 0, 0, 0, 0, 0, 0, 3);
+        assert!(Ipv6Range::new(lo, hi).is_ok());
+        assert!(Ipv6Range::new(lo, lo).is_ok());
+        assert!(Ipv6Range::new(hi, lo).is_err());
+    }
+
+    #[test]
+    fn test_ip_range_enum_deserialization() {
+        let data = r#"{"V4": {"first": "10.0.0.1", "last": "10.0.0.3"}}"#;
+        let expected = IpRange::V4(
+            Ipv4Range::new(
+                Ipv4Addr::new(10, 0, 0, 1),
+                Ipv4Addr::new(10, 0, 0, 3),
+            )
+            .unwrap(),
+        );
+        assert_eq!(expected, serde_json::from_str(data).unwrap());
+
+        let data = r#"{"V6": {"first": "fd00::", "last": "fd00::3"}}"#;
+        let expected = IpRange::V6(
+            Ipv6Range::new(
+                Ipv6Addr::new(0xfd00, 0, 0, 0, 0, 0, 0, 0),
+                Ipv6Addr::new(0xfd00, 0, 0, 0, 0, 0, 0, 3),
+            )
+            .unwrap(),
+        );
+        assert_eq!(expected, serde_json::from_str(data).unwrap());
+
+        let data = r#"{"V6": {"first": "fd00::3", "last": "fd00::"}}"#;
+        assert!(
+            serde_json::from_str::<IpRange>(data).is_err(),
+            "Expected an error deserializing an IP range with first address \
+            greater than last address",
+        );
+    }
+
+    #[test]
+    fn test_ip_range_try_from() {
+        let lo = Ipv4Addr::new(10, 0, 0, 1);
+        let hi = Ipv4Addr::new(10, 0, 0, 3);
+        assert!(IpRange::try_from((lo, hi)).is_ok());
+        assert!(IpRange::try_from((hi, lo)).is_err());
+
+        let lo = Ipv6Addr::new(0xfd00, 0, 0, 0, 0, 0, 0, 1);
+        let hi = Ipv6Addr::new(0xfd00, 0, 0, 0, 0, 0, 0, 3);
+        assert!(IpRange::try_from((lo, hi)).is_ok());
+        assert!(IpRange::try_from((hi, lo)).is_err());
     }
 }
