+++ nexus/src/external_api/tag-config.json
@@ -39,7 +39,7 @@
       }
     },
     "ip-pools": {
-      "description": "IP Pools contain subnets used for assigning external IP addresses to virtual machine instances.",
+      "description": "IP Pools contain external IP addresses that can be assigned to virtual machine Instances.",
       "external_docs": {
         "url": "http://oxide.computer/docs/#xxx"
       }
